consumer-producer/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.3
MarkupSafe==1.1.0
Werkzeug==0.14.1
